fix: use BigDecimal for monetary values

+++ src/main/java/com/trading/platform/dto/OrderResponse.java
@@ -1,10 +1,12 @@
 package com.trading.platform.dto;
 
+import java.math.BigDecimal;
+
 public record OrderResponse(
         Long id,
         Long productId,
         String productName,
         Integer quantity,
-        double totalPrice
+        BigDecimal totalPrice
 ) {
 }
